requirements.txt
fastapi
uvicorn
pandas
streamlit
requests
plotly
scipy
python-multipart
psycopg2-binary
sqlalchemy
mccabe
python-dotenv